+++ src/main/java/com/github/theredbrain/mergeditems/config/ServerConfig.java
@@ -3,7 +3,12 @@ package com.github.theredbrain.mergeditems.config;
 import com.github.theredbrain.mergeditems.MergedItems;
 import me.fzzyhmstrs.fzzy_config.annotations.Comment;
 import me.fzzyhmstrs.fzzy_config.config.Config;
+import me.fzzyhmstrs.fzzy_config.validation.minecraft.ValidatedIdentifier;
 import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedBoolean;
+import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedDouble;
+import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedInt;
+import net.minecraft.registry.Registries;
+import net.minecraft.util.Identifier;
 
 public class ServerConfig extends Config {
 
@@ -17,5 +22,10 @@ public class ServerConfig extends Config {
 				When set to 'true', similar modifiers (same id, same slot, same operation) will be averaged.
 				""")
 	public ValidatedBoolean merging_averages_similar_modifiers = new ValidatedBoolean(true);
+	public ValidatedInt default_item_cost_amount = new ValidatedInt(0);
+	public ValidatedDouble default_merging_item_cost_multiplier = new ValidatedDouble(1.0);
+	public ValidatedDouble default_splitting_item_cost_multiplier = new ValidatedDouble(1.0);
+	public ValidatedInt default_merged_items_amount_maximum = new ValidatedInt(1);
+	public ValidatedIdentifier merging_item_cost = ValidatedIdentifier.ofRegistry(Identifier.of("minecraft:lapis_lazuli"), Registries.ITEM);
 
 }