+++ src/main/java/com/github/theredbrain/mergeditems/network/packet/OpenItemMergingScreenPacketReceiver.java
@@ -18,14 +18,17 @@ public class OpenItemMergingScreenPacketReceiver implements ServerPlayNetworking
 	@Override
 	public void receive(OpenItemMergingScreenPacket payload, ServerPlayNetworking.Context context) {
 
-		int maxMergedItemsAmount = payload.maxMergedItemsAmount();
+		int defaultItemCostAmount = payload.defaultItemCostAmount();
+		double mergingItemCostMultiplier = payload.mergingItemCostMultiplier();
+		double splittingItemCostMultiplier = payload.splittingItemCostMultiplier();
+		int mergedItemsAmountMaximum = payload.mergedItemsAmountMaximum();
 		String title = payload.title();
 		List<String> list = payload.list();
 
 		context.player().openHandledScreen(new ExtendedScreenHandlerFactory<>() {
 			@Override
 			public ItemMergingScreenHandler.ItemMergingData getScreenOpeningData(ServerPlayerEntity player) {
-				return new ItemMergingScreenHandler.ItemMergingData(maxMergedItemsAmount, list);
+				return new ItemMergingScreenHandler.ItemMergingData(defaultItemCostAmount, mergingItemCostMultiplier, splittingItemCostMultiplier, mergedItemsAmountMaximum, list);
 			}
 
 			@Override
@@ -36,7 +39,7 @@ public class OpenItemMergingScreenPacketReceiver implements ServerPlayNetworking
 			@Nullable
 			@Override
 			public ScreenHandler createMenu(int syncId, PlayerInventory playerInventory, PlayerEntity player) {
-				return new ItemMergingScreenHandler(syncId, playerInventory, maxMergedItemsAmount, list);
+				return new ItemMergingScreenHandler(syncId, playerInventory, defaultItemCostAmount, mergingItemCostMultiplier, splittingItemCostMultiplier, mergedItemsAmountMaximum, list);
 			}
 		});
 	}